move getFilesByTypeAndProcessFields back and fix up tests.
Had to move it back for the time being, since including it in handleFieldsJs was causing issues with the tests
